api/requirements.txt
fastapi
uvicorn
python-dotenv
openai
psycopg2-binary
requests
redis